[TM-3592] resolved comments
